refactor: move the classes related to the robot into its own package
This separates the code that the students should interact with into
a single package
