+++ src/no/itera/lego/util/EV3Helper.java
@@ -4,6 +4,7 @@ import lejos.hardware.Sound;
 import lejos.hardware.motor.EV3LargeRegulatedMotor;
 import lejos.hardware.port.MotorPort;
 import lejos.robotics.RegulatedMotor;
+import no.itera.lego.robot.RobotController;
 
 /**
  * Helper functions for basic operation of our pre-configured robot.
